backend/requirements.txt
aiofiles==0.8.0
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.0
asyncpg==0.25.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
Click==7.0
colorama==0.4.3
cryptography==36.0.1
databases==0.5.3
distlib==0.3.4
fastapi==0.65.2
filelock==3.4.0
greenlet==1.1.2
gunicorn==20.0.4
h11==0.9.0
httptools==0.3.0
humps==0.2.2
idna==3.3
isort==4.3.21
lazy-object-proxy==1.4.3
marshmallow==3.14.1
mccabe==0.6.1
pipenv==2021.11.23
pkg_resources==0.0.0
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.6.2
PyJWT==2.3.0
pylint==2.12.2
python-env-loader==0.0.2
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.14.2
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.12.0
uvloop==0.14.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
websockets==10.1
wrapt==1.11.2
